Entrega de Proyecto

+++ EcommerceAPI.Comunes/Classes/Contracts/Ecommerce/ProductoContract.cs
@@ -8,9 +8,9 @@ namespace EcommerceAPI.Comunes.Classes.Contracts.Ecommerce
 {
     public class ProductoContract
     {
-        public int idProducto { get; set; }
+        public Int32 idProducto { get; set; }
         public string descripcion { get; set; }
-        public int stock { get; set; }
+        public decimal stock { get; set; }
         public decimal ValorUnitario { get; set; }
         public int Estado { get; set; }
         public string Imagen { get; set; }
