+++ EcommerceAPI.Infraestructura/Database/Entities/ProductoEntity.cs
@@ -7,9 +7,9 @@ namespace EcommerceAPI.Infraestructura.Database.Entities
     public class ProductoEntity
     {
         [Key]
-        public int idProducto {  get; set; }
+        public Int32 idProducto {  get; set; }
         public string descripcion { get; set; }
-        public int stock { get; set; }
+        public decimal stock { get; set; }
         public decimal ValorUnitario { get; set; }
         public int Estado { get; set; } 
         public string Imagen { get; set; }
